move provenance.py to common
