+++ ebl/corpus/application/schemas.py
@@ -1,5 +1,5 @@
 from ebl.corpus.domain.manuscript_type import ManuscriptType
-from ebl.corpus.domain.provenance import Provenance
+from ebl.common.domain.provenance import Provenance
 from marshmallow import (
     Schema,
     ValidationError,
