+++ ebl/fragmentarium/domain/findspot.py
@@ -3,7 +3,7 @@ from typing import Optional, Sequence
 from enum import Enum, auto
 from ebl.bibliography.domain.reference import Reference
 from ebl.fragmentarium.domain.date_range import DateRange
-from ebl.corpus.domain.provenance import Provenance
+from ebl.common.domain.provenance import Provenance
 
 
 ExcavationSite = Provenance
