+++ ebl/tests/corpus/test_chapter.py
@@ -16,7 +16,7 @@ from ebl.corpus.domain.manuscript import (
     Siglum,
 )
 from ebl.corpus.domain.manuscript_type import ManuscriptType
-from ebl.corpus.domain.provenance import Provenance
+from ebl.common.domain.provenance import Provenance
 from ebl.corpus.domain.record import Author, AuthorRole, Record, Translator
 from ebl.transliteration.domain.stage import Stage
 from ebl.transliteration.domain.text_id import TextId
