+++ ebl/tests/corpus/test_manuscript.py
@@ -5,7 +5,7 @@ from ebl.common.domain.period import Period
 
 from ebl.corpus.domain.manuscript import Manuscript
 from ebl.corpus.domain.manuscript_type import ManuscriptType
-from ebl.corpus.domain.provenance import Provenance
+from ebl.common.domain.provenance import Provenance
 from ebl.tests.factories.corpus import ManuscriptLineFactory
 from ebl.transliteration.domain.line import EmptyLine
 from ebl.transliteration.domain.line_number import LineNumber
